requirements.txt
appdirs==1.4.4
asgiref==3.2.7
certifi==2020.4.5.2
chardet==3.0.4
cx-Oracle==7.3.0
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-allauth==0.42.0
django-debug-toolbar==2.2
django-extensions==2.2.9
django-sslserver==0.22
django-widget-tweaks==1.4.8
filelock==3.0.12
html5lib==1.1
idna==2.9
importlib-metadata==1.6.0
oauthlib==3.1.0
path==14.0.1
pathlib==1.0.1
Pillow==7.1.2
PyPDF2==1.26.0
python3-openid==3.1.0
pytz==2019.3
repath==0.9.0
reportlab==3.5.42
requests==2.24.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.9
virtualenv==20.0.20
webencodings==0.5.1
xhtml2pdf==0.2.4
zipp==3.1.0
